requirements.txt
pypdf
requests
mammoth
pypandoc_binary
PyGithub
pypandoc
pdfminer
pdfminer.six